Slight reordering of files and includes
