+++ src/proximalMaps/ProximalMaps.jl
@@ -1,0 +1,11 @@
+include("ProxL1.jl")
+include("ProxL2.jl")
+include("ProxL21.jl")
+include("ProxLLR.jl")
+# includes/ProxSLR.jl")
+include("ProxPositive.jl")
+include("ProxProj.jl")
+include("ProxReal.jl")
+include("ProxTV.jl")
+include("ProxTVCondat.jl")
+include("ProxNuclear.jl")
